+++ core/src/reader/xml_reader/string_event_handler.rs
@@ -57,7 +57,7 @@ impl EventHandler for StringEventHandler {
         &self,
         _tag_name: String,
         _attributes: Vec<OwnedAttribute>,
-    ) -> Result<Box<EventHandler>, InnerError> {
+    ) -> Result<Box<dyn EventHandler>, InnerError> {
         Ok(Box::new(SinkingEventHandler::new()))
     }