+++ test_utilities/src/res.rs
@@ -1,4 +1,3 @@
-use std::fmt::Debug;
 use std::fs;
 use std::fs::File;
 use std::path::{Path, PathBuf};
@@ -21,7 +20,7 @@ pub fn setup_values_dir_for_locale<P: AsRef<Path>, L: AsRef<str>>(
 
 pub fn setup_empty_strings_for_default_locale<P: AsRef<Path>>(res_path: P) -> FileWithPath {
     let values_dir_path = setup_values_dir_for_default_locale(res_path);
-    setup_strings_file::<String, String>(values_dir_path, None)
+    setup_strings_file::<String, String>(values_dir_path)
 }
 
 pub fn setup_empty_strings_for_locale<P: AsRef<Path>, L: AsRef<str>>(
@@ -29,7 +28,7 @@ pub fn setup_empty_strings_for_locale<P: AsRef<Path>, L: AsRef<str>>(
     locale_id: L,
 ) -> FileWithPath {
     let values_dir_path = setup_values_dir_for_locale(res_path, locale_id.as_ref().clone());
-    setup_strings_file::<String, L>(values_dir_path, Some(locale_id))
+    setup_strings_file::<String, L>(values_dir_path)
 }
 
 fn setup_values_dir<P: AsRef<Path>, L: AsRef<str>>(res_path: P, locale_id: Option<L>) -> String {
@@ -45,7 +44,6 @@ fn setup_values_dir<P: AsRef<Path>, L: AsRef<str>>(res_path: P, locale_id: Optio
 
 fn setup_strings_file<P: AsRef<Path>, L: AsRef<str>>(
     locale_values_dir_path: P,
-    locale_id: Option<L>,
 ) -> FileWithPath {
     let mut strings_file_path = PathBuf::from(locale_values_dir_path.as_ref());
     strings_file_path.push("strings.xml");