+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.1.8
+- Add new optional argument "strict" to the validate command
+  - if present, will also validate that all strings have been localized
+
 # 0.1.7
 - Update Cargo.toml to say 0.1.7 too
 - Carry along the extension for Windows binary